requirements.txt
moviepy==1.0.3
opencv-python==4.5.3.56
Pillow==8.3.1

numpy==1.21.1
tk==0.1.0
segment-anything==0.1.0